restoring load to running functionality
